flask_rested/requirements.txt
aniso8601==9.0.1
blinker==1.5
click==8.1.3
commonmark==0.9.1
Flask==2.2.2
Flask-RESTful==0.3.9
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
Pygments==2.13.0
python-dotenv==0.20.0
pytz==2022.2.1
rich==12.5.1
six==1.16.0
typing_extensions==4.3.0
Werkzeug==2.2.2
zipp==3.8.1
